requirements.txt
autopep8==1.3.2
click==6.7
Django==1.11.3
django-extensions==1.8.1
djangorestframework==3.6.3
drf-nested-routers==0.90.0
nose==1.3.7
numpy==1.11.2
psycopg2==2.7.1
pycodestyle==2.3.1
python-dotenv==0.6.4
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
